fix some bugs

+++ src/main/java/Models/StatsModels/EndToEndDelay.java
@@ -2,15 +2,15 @@ package Models.StatsModels;
 
 public class EndToEndDelay {
 
-    private long totalDelay = 0;
+    private float totalDelay = 0;
     private int packetCount = 0;
 
-    public void addPacket(long delay){
+    public void addPacket(float delay) {
         totalDelay += delay;
         packetCount++;
     }
 
-    public EndToEndDelay merge(EndToEndDelay endToEndDelay){
+    public EndToEndDelay merge(EndToEndDelay endToEndDelay) {
         EndToEndDelay temp = new EndToEndDelay();
         temp.totalDelay += endToEndDelay.totalDelay;
         temp.totalDelay += this.totalDelay;
@@ -19,7 +19,7 @@ public class EndToEndDelay {
         return temp;
     }
 
-    public float getAverage(){
-        return (float) totalDelay / (float) packetCount;
+    public float getAverage() {
+        return totalDelay / (float) packetCount;
     }
 }
